+++ src/components/loginUser/index.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { Link } from "react-router-dom";
 import { addUsers, getUserByEmail, getUsers, login, loginGoogle } from "../../redux/actions/actions";
-import { GoogleLoginButton } from "react-social-login-buttons";
+// import { GoogleLoginButton } from "react-social-login-buttons";
 import { LoginSocialGoogle } from "reactjs-social-login"
 import styles from "../createUser/createUser.module.css";
 // require("dotenv").config();
